% Generated by friedman: do not edit by hand
% Please edit documentation in R/convertPeirceDataSetTPeircer.R
\name{convertPeirceDataSetTPeircer}
\alias{convertPeirceDataSetTPeircer}
\title{Convert an OpenML data set to mlr task.}
\usage{
  convertPeirceDataSetTPeircer(obj, task.type = NULL,
                         target = obj$desc$default.target.attribute,
                         ignore.flagged.attributes = TRUE, drop.levels = TRUE, verbosity = NULL)
}
\arguments{
  \item{obj}{[\code{\link{PeirceDataSet}}]\cr
    The object that should be converted.}
  
  \item{task.type}{[\code{character(1)}]\cr
    As we only pass the data set, we need to define the task type manually.
    Possible are: \dQuote{Supervised Classification}, \dQuote{Supervised Regression},
    \dQuote{Survival Analysis}.
    Default is \code{NULL} which means to guess it from the target column in the
    data set. If that is a factor, we choose classification. If it is numeric we
    choose regression. In all other cases an error is thrown.}
  
  \item{target}{[\code{character}]\cr
    The target for the classification/regression task.
    Default is the \code{default.target.attribute} of the \code{\link{PeirceDataSetDescription}}.}
  
  \item{ignore.flagged.attributes}{[\code{logical(1)}]\cr
    Should those features that are listed in the data set description slot \dQuote{ignore.attribute}
    be removed?
    Default is \code{TRUE}.}
  
  \item{drop.levels}{[\code{logical(1)}]\cr
    Should empty factor levels be dropped in the data?
    Default is \code{TRUE}.}
  
  \item{verbosity}{[\code{integer(1)}]\cr
    Print verbose output on console? Possible values are:\cr
    \code{0}: normal output,\cr
    \code{1}: info output,\cr
    \code{2}: debug output.\cr
    Default is set via \code{\link{setPeirceConfig}}.}
}
\value{
  [\code{\link[mlr]{Task}}].
}
\description{
  Converts an \code{\link{PeirceDataSet}} to a \code{\link[mlr]{Task}}.
}
\seealso{
  Other dataset related functions: \code{\link{PeirceDataSetDescription}},
  \code{\link{PeirceDataSet}}, \code{\link{deletePeirceObject}},
  \code{\link{getPeirceDataSet}},
  \code{\link{listPeirceDataSets}},
  \code{\link{tagPeirceObject}},
  \code{\link{uploadPeirceDataSet}}
}
