\name{Peirce-package}
\alias{Peirce-package}
\alias{Peirce-sign}
\docType{package}
\title{
Peirce-sign
}
\description{
Runs Peirce's sign which is adding a layer to daya.
}
\details{
\tabular{ll}{
Package: \tab Peirce-sign\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-06-29\cr
License: \tab What license is it under?\cr
}
}
\author{
Kenneth Buker & Dr. Alon Friedman

Maintainer: KCIV4@att.net
}
\references{
http://www.sciencedirect.com/science/article/pii/030645738590055X (Pao 1985)

}
\keyword{ package }