\name{Peirce}
\alias{Peice}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Runs the creation of object classification value 
}
\description{
Generates Peirce classification
\usage{
LotkasC(N)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{N}{
Value N Generated by PeirceN
}
}

\value{
Value Returned is C.
}

\author{
Alon Friedman
}


%% ~Make other sections like Warning with \section{Warning }{....} ~


\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

